% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connICA.R
\name{connICA}
\alias{connICA}
\title{connICA}
\usage{
connICA(data, q)
}
\arguments{
\item{data}{Data matrix}

\item{q}{number of sources}
}
\value{
a R list containing the following terms:
\describe{
  \item{A}{the estimated mixing coefficent matrix.}
  \item{S}{the estimated latent source matrix.}
}
}
\description{
connICA
}
