% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSS_gibbs_sampler_summary.R
\name{DLSS_gibbs_sampler_summary}
\alias{DLSS_gibbs_sampler_summary}
\title{Title}
\usage{
DLSS_gibbs_sampler_summary(res)
}
\arguments{
\item{res}{result from the function NLSS}
}
\description{
Title
}
