% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_comp.R
\name{match_comp}
\alias{match_comp}
\title{Title}
\usage{
match_comp(Slist)
}
\arguments{
\item{Slist}{list of q*p discrete value source matrics}
}
\value{
list of q*p discrete value source matrics with rows matched to  the first matrix
}
\description{
Title
}
