% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLSS.R
\name{NLSS}
\alias{NLSS}
\title{MCMC Sampling for the NLSS Model
This function performs Markov chain Monte Carlo (MCMC) sampling for the
Network Latent Source Separation (NLSS) model. Given a discrete-valued
network data matrix, the function estimates the latent source matrix, the
mixing coefficient matrix, and related posterior samples.}
\usage{
NLSS(
  data,
  q = 2,
  q0 = q,
  init = list(S = NULL, A = NULL, Y = NULL),
  states = (min(data):max(data)),
  state0 = min(data),
  total_iter = 1000,
  burn_in = 500,
  thin = 10,
  show_step = 100,
  joint = TRUE
)
}
\arguments{
\item{data}{An \eqn{n \times p} matrix of discrete-valued connection states,
where \eqn{n} is the sample size and \eqn{p} is the number of node pairs.}

\item{q}{An integer specifying the number of latent sources.}

\item{q0}{An integer specifying the maximum number of non-background sources
allowed for each node pair. The default is \code{q}.}

\item{init}{A list of initial values for the MCMC sampler. It may contain
\code{S}, \code{A}, and \code{Y}, corresponding to the initial latent source
matrix, mixing coefficient matrix, and latent allocation matrix,
respectively. Missing components are initialized automatically.}

\item{states}{A vector containing all possible discrete states in the observed
data and latent sources. The default is \code{min(data):max(data)}.}

\item{state0}{The background or null state. The default is \code{min(data)}.}

\item{total_iter}{An integer specifying the total number of MCMC iterations.
The default is \code{1000}.}

\item{burn_in}{An integer specifying the number of initial MCMC iterations to
be discarded as burn-in. The default is \code{500}.}

\item{thin}{An integer specifying the thinning interval for saving posterior
samples. The default is \code{10}.}

\item{show_step}{An integer specifying how often the current iteration number
is printed during MCMC sampling. The default is \code{100}.}

\item{joint}{A logical value indicating whether joint updating is used in the
MCMC sampler. The default is \code{TRUE}.}
}
\value{
A list containing posterior samples and model information from the
  MCMC sampler. The returned object includes, but is not necessarily limited
  to, the following components:
\describe{
  \item{\code{A}}{Posterior samples of the mixing coefficient matrix.}
  \item{\code{S}}{Posterior samples of the latent source matrix.}
  \item{\code{Y}}{Posterior samples of the latent allocation matrix.}
  \item{\code{X}}{The input discrete-valued data matrix.}
  \item{\code{K}}{The number of possible discrete states.}
  \item{\code{states}}{The vector of possible discrete states.}
  \item{\code{state0}}{The background or null state.}
}
}
\description{
MCMC Sampling for the NLSS Model
This function performs Markov chain Monte Carlo (MCMC) sampling for the
Network Latent Source Separation (NLSS) model. Given a discrete-valued
network data matrix, the function estimates the latent source matrix, the
mixing coefficient matrix, and related posterior samples.
}
