% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{sim_NLSS}
\alias{sim_NLSS}
\title{Simulate data from a NLSS model adding Gaussian noise}
\usage{
sim_NLSS(n = 50, alpha_0 = c(0.5, 0.5, 0.5), alpha_1 = 0.1, sd0 = 0.3, S)
}
\arguments{
\item{n}{sample size}

\item{alpha_0}{concentration parameter (a three dimensional vector) for latent sources.}

\item{alpha_1}{concentration parameter (a scale variable) for noise.}

\item{sd0}{standard deviation for the Gaussian noise}

\item{S}{the latent source matrix}
}
\value{
a R list containing the following terms:
\describe{
  \item{Xc}{the data matrix (with Gaussian noise)}
  \item{X}{the data matrix (without Gaussian noise)}
  \item{A}{the mixing coefficient matrix}
}
}
\description{
This function simulates observations using NLSS with given latent sources.
}
