% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{gen_sources}
\alias{gen_sources}
\title{Generate latent sources}
\usage{
gen_sources(n_node = 50)
}
\arguments{
\item{n_node}{number of nodes.}
}
\value{
a R list containing the following terms:
\describe{
  \item{S}{the latent source matrix with each row representing a source network (vectorized adjacency matrix).}
  \item{community}{the community index for each node.}
}
}
\description{
This function simulates generate three latent source networks with eight communities.
}
