% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heattmap.net.R
\name{heatmap.net}
\alias{heatmap.net}
\title{Heatmaps for Networks}
\usage{
heatmap.net(
  S,
  lim = c(min(S), max(S)),
  community = rep(1, (1 + sqrt(1 + 8 * ncol(S)))/2),
  color = bluered(100)
)
}
\arguments{
\item{S}{a matrix with each row representing a vectorized weighted adjacency matrix.}

\item{lim}{a 2-dimentional vector specifying the limits for the data.}

\item{community}{a vector represents the community each node belongs to.}

\item{color}{colorbar for the heatmap.}
}
\description{
The function plots heatmaps for single or multiple network(s).
}
