% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_comp.R
\name{match_rows}
\alias{match_rows}
\title{Match discrete value matrices by their rows}
\usage{
match_rows(Slist)
}
\arguments{
\item{Slist}{list of discrete value source matrices with same dimensions.}
}
\value{
A list of discrete value source matrices with rows matched to the rows of the first matrix.
}
\description{
The function reorders the rows of matrices to match them to rows of the
first matrix in the list.
}
