% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLSS.R
\name{NLSS}
\alias{NLSS}
\title{MCMC sampling for NLSS}
\usage{
NLSS(
  data,
  q = 2,
  q0 = 2,
  init = list(S = NULL, A = NULL, Y = NULL),
  states = (min(data):max(data)),
  state0 = min(data),
  total_iter = 1000,
  burn_in = 500,
  thin = 10,
  show_step = 100,
  joint = TRUE
)
}
\arguments{
\item{data}{a n*p matrix with discrete-valued connection states, n is the sample size, p is the num of node pairs.}

\item{q}{the number of latent sources.}

\item{init}{a list of initial values for S and A.}

\item{states}{a vector of all possible states.}

\item{state0}{the zero states.}

\item{total_iter}{the number of total iterations.}

\item{burn_in}{the number of iterations to be discarded as burn-in.}

\item{show_step}{the frequency for printing the current number of iterations.}
}
\value{
a R list containing the following terms:
\describe{
  \item{A}{the estimated mixing coefficent matrix.}
  \item{beta_coef}{the frequency of each discrete value of S appearing
in the MCMC samples.}
  \item{S}{the estimated latent source matrix.}
  \item{logLik}{the log-likelihood trace.}
}
}
\description{
The function conducts MCMC sampling for the NLSS model.
}
