% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLSS.R
\name{NLSS}
\alias{NLSS}
\title{MCMC sampling for NLSS}
\usage{
NLSS(data, q = 2, total_iter = 1000, burn_in = 500,
  show_step = 100)
}
\arguments{
\item{data}{a n*p matrix with discrete value entries, n is the sample size, p is the num of node pairs}

\item{q}{the number of latent sources}

\item{total_iter}{the number of total iterations}

\item{burn_in}{the number of iterations to be discarded as burn-in}

\item{show_step}{the number of iterations as a gap between two consecutive prints for monitoring the algorithm}
}
\value{

}
\description{
The function conducts MCMC sampling for a NLSS model. The number of latent sources need to be specified.
}
\details{
f
}
