% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relia_comp.R
\name{relia_rows}
\alias{relia_rows}
\title{Reliability of sources}
\usage{
relia_rows(Slist)
}
\arguments{
\item{Slist}{a list of matrices with rows representing vectorization of latent sources.}
}
\value{
Reliability index.
}
\description{
The function calculates the reliability index of sources for NLSS. The input is a list of
matrices with the first matrix being estimated sources from the whole dataset and the following matrices
being estimated sources from bootsrap samples of the whole dataset.
}
