% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{sim_ICA}
\alias{sim_ICA}
\title{Simulate data from a linear mixture with Gaussian noise}
\usage{
sim_ICA(n = 50, sd = c(0.5, 0.5, 0.5), sd0 = 0.1, S)
}
\arguments{
\item{n}{sample size}

\item{sd}{standard deviation for the Gaussian mixing coefficient}

\item{sd0}{standard deviation for the Gaussian noise}

\item{S}{the latent source matrix}
}
\value{
a R list containing the following terms:
\describe{
  \item{X}{the data matrix}
  \item{A}{the mixing coefficient matrix}
  \item{noise}{the noise matrix}
}
}
\description{
This function simulates observations using a linear mixture with given latent sources.
}
