% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{sim_NLSS}
\alias{sim_NLSS}
\title{Simulate Data from an NLSS Model with Gaussian Noise}
\usage{
sim_NLSS(n = 50, alpha_0 = c(0.5, 0.5, 0.5), alpha_1 = 0.1, sd0 = 0.3, S)
}
\arguments{
\item{n}{An integer specifying the sample size. The default is \code{50}.}

\item{alpha_0}{A numeric vector of Dirichlet concentration parameters for the
latent source components. Its length should be equal to the number of
latent sources, namely \code{nrow(S)}. The default is
\code{c(0.5, 0.5, 0.5)}.}

\item{alpha_1}{A numeric value specifying the Dirichlet concentration
parameter for the background or noise component. The default is \code{0.1}.}

\item{sd0}{A numeric value specifying the standard deviation of the additive
Gaussian noise. The default is \code{0.3}.}

\item{S}{A latent source matrix, where each row represents one source network
in vectorized adjacency matrix form.}
}
\value{
A list containing the following components:
\describe{
  \item{\code{Xc}}{The simulated data matrix with additive Gaussian noise.}
  \item{\code{X}}{The noiseless simulated data matrix generated from the
  NLSS model.}
  \item{\code{A}}{The simulated mixing coefficient matrix.}
}
}
\description{
This function simulates observations from the Network Latent Source
Separation (NLSS) model with given latent source networks. Mixing coefficients
are generated from a Dirichlet distribution, and optional Gaussian noise is
added to the noiseless NLSS observations.
}
