% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{sim_NLSS}
\alias{sim_NLSS}
\title{simulate data from a NLSS model}
\usage{
sim_NLSS(n_node = 50, n = 40, alpha_0 = c(0.5, 0.5, 0.5), alpha_1 = 0.1)
}
\arguments{
\item{n_node}{number of nodes.}

\item{n}{sample size.}

\item{alpha}{relative signal strengh.}

\item{beta}{relative noise strengh.}
}
\value{
a R list containing the following terms:
\describe{
  \item{X}{the observed matrix.}
  \item{S}{the latent source matrix.}
  \item{A}{the mixing coefficent matrix.}
  \item{community}{the community index.}
}
}
\description{
This function simulates observations using NLSS with pre-specified latent sources.
The number of latent sources is three, with one representing within-community connections and two
representing cross-comminity connections.
}
