% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_NLSS.R
\name{sim_NLSS}
\alias{sim_NLSS}
\title{Title}
\usage{
sim_NLSS(n_node = 50, n = 40, alpha = 0.5, beta = 0.1)
}
\arguments{
\item{n_node}{number of nodes}

\item{n}{sample size}

\item{alpha}{alpha}

\item{beta}{beta}
}
\description{
Title
}
