% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_mat.R
\name{vec_mat}
\alias{vec_mat}
\title{transformation between a symmetric matrix and a vector}
\usage{
vec_mat(x)
}
\arguments{
\item{x}{a symmetric matrix or a vector}
}
\value{
a vector or a symmetric matrix
}
\description{
\code{vec_mat} is a small function that vectorize a symmetric
matrix (only keep the upper triangle) or transform a vector to a symmetric
matrix
}
\examples{
x = matrix(rnorm(36),nrow = 6)
vec_mat(x)
}
