% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_matlist.R
\name{vec_matlist}
\alias{vec_matlist}
\title{Vectrization of a list of symmetric matrices}
\usage{
vec_matlist(mat_list, q = 0, binarize = FALSE)
}
\arguments{
\item{mat_list}{a list of matrices}

\item{q}{a quantile serving as a threshold, any value smaller than the quantile will be set to zero}

\item{binarize}{binarize the outcome or not}
}
\value{
a matrix each row representing a matrix from the input mat_list
}
\description{
Vectorize a list of symmetric matrices and stack them together.
The outcome can be thresholded and binarized with optional parameters.
}
